Make algorithm in nOrMoreInCell more efficient
The old algorithm was efficient if there were only categorical values
but scaled quadratically if the rows were all different. For example
when there is a continuous variable. The new algorithm is a bit of a
hack because I would prefer to calculate a hash for each row, but the
paste0(..., collapse) works similar and is only slight less memory
efficient. However, the new algorithm doesn't have any of the scaling
problems